seperated methods into more interfaces
