+++ felt-item-api/src/main/java/io/github/feltmc/feltapi/api/item/extensions/ContainerItemExtension.java
@@ -22,4 +22,8 @@ public interface ContainerItemExtension {
     {
         return self().hasRecipeRemainder();
     }
+
+    default boolean isRepairable(){
+        return false;
+    }
 }
