+++ felt-item-api/src/main/java/io/github/feltmc/feltapi/mixin/item/LivingEntityMixin.java
@@ -4,6 +4,8 @@ import io.github.feltmc.feltapi.api.item.ToolActions;
 import io.github.feltmc.feltapi.api.item.event.LivingEntityUsageTickCallback;
 import io.github.feltmc.feltapi.api.item.extensions.ArmorExtension;
 import io.github.feltmc.feltapi.api.item.extensions.FeltItem;
+import io.github.feltmc.feltapi.api.item.extensions.ItemUseExtension;
+import io.github.feltmc.feltapi.api.item.extensions.MiscExtension;
 import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.item.Item;
@@ -47,7 +49,7 @@ public abstract class LivingEntityMixin {
 
     @Redirect(method = "getPreferredEquipmentSlot", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
     private static boolean redirectGetEquipmentSlot(ItemStack instance, Item item){
-        if (instance.getItem() instanceof FeltItem feltItem) {
+        if (instance.getItem() instanceof MiscExtension feltItem) {
             return feltItem.canPerformAction(instance, ToolActions.SHIELD_BLOCK);
         }
         return instance.isOf(item);
@@ -63,7 +65,7 @@ public abstract class LivingEntityMixin {
     private void tickItemStackUsage(CallbackInfo ci) {
         if (!this.activeItemStack.isEmpty()) {
             itemUseTimeLeft = LivingEntityUsageTickCallback.TICK.invoker().tick((LivingEntity)(Object)this, activeItemStack, itemUseTimeLeft);
-            if (itemUseTimeLeft > 0 && activeItemStack.getItem() instanceof FeltItem item)
+            if (itemUseTimeLeft > 0 && activeItemStack.getItem() instanceof ItemUseExtension item)
                 item.onUsingTick(activeItemStack, (LivingEntity) (Object)this, itemUseTimeLeft);
         }
 
@@ -73,7 +75,7 @@ public abstract class LivingEntityMixin {
     {
         if (!from.isEmpty() && !to.isEmpty())
         {
-            return from.getItem() instanceof FeltItem item && item.canContinueUsing(from, to);
+            return from.getItem() instanceof ItemUseExtension item && item.canContinueUsing(from, to);
         }
         return false;
     }
