+++ felt-item-api/src/main/java/io/github/feltmc/feltapi/mixin/item/PiglinEntityMixin.java
@@ -1,6 +1,7 @@
 package io.github.feltmc.feltapi.mixin.item;
 
 import io.github.feltmc.feltapi.api.item.extensions.FeltItem;
+import io.github.feltmc.feltapi.api.item.extensions.PiglinExtension;
 import net.minecraft.entity.mob.PiglinEntity;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
@@ -13,7 +14,7 @@ public class PiglinEntityMixin {
 
     @Redirect(method = "equipToOffHand", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
     private boolean redirectPiglinItem(ItemStack instance, Item item){
-        if (instance.getItem() instanceof FeltItem item1){
+        if (instance.getItem() instanceof PiglinExtension item1){
             return item1.isPiglinCurrency(instance);
         }
         return instance.isOf(item);
