+++ felt-item-api/src/main/java/io/github/feltmc/feltapi/mixin/item/ServerPlayerEntityMixin.java
@@ -2,6 +2,7 @@ package io.github.feltmc.feltapi.mixin.item;
 
 import com.mojang.authlib.GameProfile;
 import io.github.feltmc.feltapi.api.item.extensions.FeltItem;
+import io.github.feltmc.feltapi.api.item.extensions.MiscExtension;
 import net.minecraft.block.BlockState;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
@@ -23,7 +24,7 @@ public abstract class ServerPlayerEntityMixin extends PlayerEntity {
     @Inject(method = "dropSelectedItem", at = @At(value = "INVOKE", target = "Lnet/minecraft/server/network/ServerPlayerEntity;getInventory()Lnet/minecraft/entity/player/PlayerInventory;", shift = At.Shift.AFTER), cancellable = true)
     private void injectDropSelectedItem(boolean entireStack, CallbackInfoReturnable<Boolean> cir){
         ItemStack selected = this.getInventory().getMainHandStack();
-        if (selected.isEmpty() || (selected.getItem() instanceof FeltItem item && !item.onDroppedByPlayer(selected,this))) {
+        if (selected.isEmpty() || (selected.getItem() instanceof MiscExtension item && !item.onDroppedByPlayer(selected,this))) {
             cir.setReturnValue(false);
         }
     }
